+++ src/lib.rs
@@ -1,29 +1,16 @@
+#![cfg_attr(not(feature = "std"), no_std)]
 #![feature(nll)]
 #![feature(external_doc)]
 #![feature(try_trait)]
 #![deny(missing_docs)]
 #![doc(include = "../README.md")]
 #![doc(html_logo_url = "https://doc.dalek.rs/assets/dalek-logo-clear.png")]
+#![doc(html_root_url = "https://docs.rs/bulletproofs/2.0.0")]
 
-extern crate byteorder;
-extern crate core;
-extern crate digest;
-extern crate rand;
-extern crate sha3;
+extern crate alloc;
 
-extern crate clear_on_drop;
-extern crate curve25519_dalek;
-extern crate merlin;
-extern crate subtle;
 #[macro_use]
 extern crate serde_derive;
-extern crate serde;
-
-#[macro_use]
-extern crate failure;
-
-#[cfg(test)]
-extern crate bincode;
 
 mod util;
 
@@ -43,17 +30,18 @@ mod inner_product_proof;
 mod range_proof;
 mod transcript;
 
-pub use errors::ProofError;
-pub use generators::{BulletproofGens, BulletproofGensShare, PedersenGens};
-pub use range_proof::RangeProof;
+pub use crate::errors::ProofError;
+pub use crate::generators::{BulletproofGens, BulletproofGensShare, PedersenGens};
+pub use crate::range_proof::RangeProof;
 
 #[doc(include = "../docs/aggregation-api.md")]
 pub mod range_proof_mpc {
-    pub use errors::MPCError;
-    pub use range_proof::dealer;
-    pub use range_proof::messages;
-    pub use range_proof::party;
+    pub use crate::errors::MPCError;
+    pub use crate::range_proof::dealer;
+    pub use crate::range_proof::messages;
+    pub use crate::range_proof::party;
 }
 
 #[cfg(feature = "yoloproofs")]
+#[cfg(feature = "std")]
 pub mod r1cs;